% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\docType{data}
\name{analysis}
\alias{analysis}
\alias{analysis_dataset}
\alias{dseval}
\alias{analysis_solver}
\alias{analysis_task}
\title{Data analysis benchmark}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 466 rows and 4 columns.
}
\usage{
analysis_dataset

analysis_solver(solver_chat = NULL)

analysis_task(epochs = 1, dir = ".vitals/logs/")
}
\description{
Objects prefixed with \code{analysis_} correspond to the data analysis benchmark in
dseval.
\itemize{
\item \code{da_dataset} is the dataset underlying the eval.
\item \code{da_solver} is the solver, which just prompts the provided chat with the
instruction and working directory provided in \code{da_dataset$input} and
allows the model to "converse" with a mock analyst—an LLM instance that
just gently prods the model along without providing any specific guidance.
\item The scorer for the data analysis task is \code{\link[vitals:scorer_model]{vitals::model_graded_qa()}}.
\item \code{da_task} combines the dataset, solver, and scorer into a \link[vitals:Task]{vitals::Task}.
}
}
\examples{
analysis_dataset

tsk <- analysis_task()
tsk

tsk$eval(solver_chat = databot:::chat_bot())

}
\keyword{datasets}
