% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dseval_dataset}
\alias{dseval_dataset}
\title{DSEval dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 466 rows and 5 columns.
}
\usage{
dseval_dataset
}
\description{
DSEval dataset
}
\keyword{datasets}
